producer/requirements.txt
kafka-python == 2.0.2
requests ==2.28.1
argparse==1.4.0
spacy==3.4.1